extract logic to reusable hook
